+++ run_tests.py
@@ -6,6 +6,7 @@ if __name__ == "__main__":
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.settings")
 
     import django
+
     django.setup()
 
     pytest.main()