+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Add official support for Django 4.0 to 4.2 with:
+- Add official support for Django 4.0 to 5.0 with:
     - Addition of automated test coverage
     - Review of all the built-in template tags and filters (no changes needed)
 - Use [tox](https://tox.wiki/) to manage test suite environment